Event loader & more abstract file checking
Moved the command loader's loadCommands() function to its own file in a more abstract form, so other modules that need to load files in a similar manner (right now for events, later probably also button-handling) can use it too without needing to duplicate the code.

That said, added an event loading module that checks the `/src/events` folder for files that export instances of the BotEvent class and has the discord.js Client instance listen to the events described therein.